Remove Config.h.in
The problem with this file is that it is used in header files, which
makes it a part of the api contract. Packaging that file with the
api headers is non-trivial.

The supplied bit functions are plenty fast, and we know that __clang__
and __GNUC__ have these functions.  If there is another platform that
has these, we can add a new define.