Remove finding the local gtest
This patch was written against v1.0 of andrewprock's pokerstove:
https://github.com/andrewprock/pokerstove
This patch fixes two upstream bugs: #73 and #74 available here:
https://github.com/andrewprock/pokerstove/issues/73
https://github.com/andrewprock/pokerstove/issues/74

Since cmake cannot detect the local gtest, the system's gtest is
required to compile correctly.
Without this patch, cmake will fail to find the local gtest during
the cmake configure phase.
After this patch, cmake will not search for the local gtest but use
the system gtest
gtest was added to the requirements listed in the README.md
gtest was removed from the source code and the cmake file

This patch was written by Lucas Mitrak and will be made available upstream.
This patch was tested locally and in a docker image created by ebuildtester