Adding swig for python brige.
Worked through much of peval library, but still need
to work through the penum library. Python support should
be considered experimental, with interface changes
expected. The final goal will be to have enough features
and packaging in place so that the Python bridge could
be installed for general use.

Additionally, the organization of CMakeLists.txt files
has bee simplified.